StorageConverterToStorageBinarySharedXXX was StorageConverterToStorageBinaryXXX
