+++ src/main/java/walkingkooka/storage/convert/StorageConverterToStorageBinarySharedExpression.java
@@ -30,18 +30,18 @@ import walkingkooka.tree.expression.Expression;
  * Converts a {@link StorageValue} into {@link StorageBinary} if the file extension is {@link FileExtension#EXPRESSION} and the value can be
  * converted into a {@link Expression} and then text and then into {@link Binary} and then {@link StorageBinary}.
  */
-final class StorageConverterToStorageBinaryExpression<C extends StorageConverterContext> extends StorageConverterToStorageBinary<C> {
+final class StorageConverterToStorageBinarySharedExpression<C extends StorageConverterContext> extends StorageConverterToStorageBinaryShared<C> {
 
     /**
      * Type safe getter.
      */
-    static <C extends StorageConverterContext> StorageConverterToStorageBinaryExpression<C> instance() {
+    static <C extends StorageConverterContext> StorageConverterToStorageBinarySharedExpression<C> instance() {
         return Cast.to(INSTANCE);
     }
 
-    private final static StorageConverterToStorageBinaryExpression INSTANCE = new StorageConverterToStorageBinaryExpression<>();
+    private final static StorageConverterToStorageBinarySharedExpression INSTANCE = new StorageConverterToStorageBinarySharedExpression<>();
 
-    private StorageConverterToStorageBinaryExpression() {
+    private StorageConverterToStorageBinarySharedExpression() {
         super();
     }
